[TAN-1836] Move user confirmation code to concern
